+++ src/transport/crypto_out_meta_sequence_transport.hpp
@@ -182,11 +182,11 @@ namespace detail {
                     char * p = mes + len;
                     *p++ = ' ';                           //     1 octet
                     for (int i = 0; i < HASH_LEN / 2; i++, p += 2) {
-                        sprintf(p, "%02x", hash[i]); //    64 octets (hash1)
+                        sprintf(p, "%02x", unsigned(hash[i])); //    64 octets (hash1)
                     }
                     *p++ = ' ';                           //     1 octet
                     for (int i = HASH_LEN / 2; i < HASH_LEN; i++, p += 2) {
-                        sprintf(p, "%02x", hash[i]); //    64 octets (hash2)
+                        sprintf(p, "%02x", unsigned(hash[i])); //    64 octets (hash2)
                     }
                     *p++ = '\n';
                     len = p-mes;
